+++ app/models/code_review.rb
@@ -5,6 +5,5 @@ class CodeReview < ApplicationRecord
   # enum status: { pending: 0, reviewed: 1, rejected: 2 } # Exemplo de uso de enum para o campo status
 
   validates :user_answer, presence: true
-  validates :received_feedback, presence: true, if: :reviewed?
-
+  validates :received_feedback, presence: true
 end
