+++ app/services/challenge_prompt_service.rb
@@ -26,7 +26,7 @@ class ChallengePromptService
           messages: [{ role: "user", content: build_challenge_prompt }], # Required.
           temperature: 0.7,
           stream: false,
-					max_tokens: 600 # might want to check this
+					max_tokens: 250 # might want to check this
       })
     # you might want to inspect the response and see what the api is giving you
     return response["choices"][0]["message"]["content"]
