+++ app/views/code_reviews/index.html.erb
@@ -1,0 +1,2 @@
+<h1>All my code reviews</h1>
+<%= link_to 'Back', root_path, class: 'btn btn-primary rounded' %>
