+++ app/views/code_reviews/show.html.erb
@@ -1,0 +1,5 @@
+<div class="container">
+<h1>Analyses and Feedback:</h1>
+<pre><%=@code_review.received_feedback %>
+</pre>
+</div>
